+++ src/services/syncProcessing/ContentDownloadService.ts
@@ -1,26 +1,32 @@
 import { Job } from 'bullmq'
+import { exec as execCallback } from 'child_process'
 import fs from 'fs'
 import fsPromises from 'fs/promises'
+import pTimeout from 'p-timeout'
 import path from 'path'
+import { promisify } from 'util'
 import { Logger } from 'winston'
 import { IDynamodbService } from '../../repository'
 import { ReadonlyConfig } from '../../types'
-import { DownloadJobData, DownloadJobOutput, YtChannel } from '../../types/youtube'
+import { DownloadJobData, DownloadJobOutput, VideoUnavailableReasons, YtChannel } from '../../types/youtube'
+import { restartEC2Instance } from '../../utils/restartEC2Instance'
 import { LoggingService } from '../logging'
 import { IYoutubeApi } from '../youtube/api'
 import { SyncUtils } from './utils'
+
+const exec = promisify(execCallback)
 
 // Youtube videos download service
 export class ContentDownloadService {
   readonly logger: Logger
 
   public constructor(
-    private syncConfig: Required<ReadonlyConfig['sync']>,
+    private config: Required<ReadonlyConfig['sync']> & ReadonlyConfig['proxy'],
     logging: LoggingService,
     private dynamodbService: IDynamodbService,
     private youtubeApi: IYoutubeApi
   ) {
-    this.syncConfig = syncConfig
+    this.config = config
     this.logger = logging.createLogger('ContentDownloadService')
     this.dynamodbService = dynamodbService
     this.youtubeApi = youtubeApi
@@ -35,7 +41,7 @@ export class ContentDownloadService {
 
   private async removeVideoFile(videoId: string) {
     try {
-      const dir = this.syncConfig.downloadsDir
+      const dir = this.config.downloadsDir
       const files = await fsPromises.readdir(dir)
       for (const file of files) {
         if (file.startsWith(videoId)) {
@@ -57,19 +63,19 @@ export class ContentDownloadService {
   }
 
   private resolveDownloadedVideos() {
-    const videoDownloadsDir = this.syncConfig.downloadsDir
+    const videoDownloadsDir = this.config.downloadsDir
     const resolvedDownloads = fs
       .readdirSync(videoDownloadsDir)
       .map((filePath) => filePath.split('.'))
       .filter((filePath) => filePath.length === 2)
       .map(([videoId, ext]) => {
-        const filePath = path.join(this.syncConfig.downloadsDir, `${videoId}.${ext}`)
+        const filePath = path.join(this.config.downloadsDir, `${videoId}.${ext}`)
         SyncUtils.setVideoFilePath(videoId, filePath)
         SyncUtils.updateUsedStorageSize(this.fileSize(filePath))
         return videoId
       })
     this.logger.verbose(`Resolved already downloaded video assets in local storage`, {
-      resolvedDownloads,
+      resolvedDownloads: resolvedDownloads.length,
       usedSpace: SyncUtils.usedSpace,
     })
   }
@@ -79,16 +85,19 @@ export class ContentDownloadService {
     const video = job.data
     try {
       // download the video from youtube
-      const { ext: fileExt } = await this.youtubeApi.downloadVideo(video.url, this.syncConfig.downloadsDir)
-      const filePath = path.join(this.syncConfig.downloadsDir, `${video.id}.${fileExt}`)
+
+      const { ext: fileExt } = await pTimeout(
+        this.youtubeApi.downloadVideo(video.url, this.config.downloadsDir),
+        this.config.limits.pendingDownloadTimeoutSec * 1000,
+        `Download timed-out`
+      )
+
+      const filePath = path.join(this.config.downloadsDir, `${video.id}.${fileExt}`)
       const size = this.fileSize(filePath)
       if (!SyncUtils.downloadedVideoFilePaths.has(video.id)) {
         SyncUtils.setVideoFilePath(video.id, filePath)
         SyncUtils.updateUsedStorageSize(size)
       }
-
-      // TODO: fix this as size can be duplicated added if video is already downloaded
-      // TODO: once during resolveDownloadedVideos calls and once during re-downloading
 
       if (video.joystreamVideo) {
         return { filePath }
@@ -111,20 +120,26 @@ export class ContentDownloadService {
       return { filePath }
     } catch (err) {
       const errorMsg = (err as Error).message
-      const errors = [
-        { message: 'Video unavailable' },
-        { message: 'Private video' },
-        { message: 'Postprocessing:' },
-        { message: 'The downloaded file is empty' },
-        { message: 'This video is private' },
-        { message: 'removed by the uploader' },
-        { message: 'size cap for historical videos' },
+      const errors: { message: string; code: VideoUnavailableReasons }[] = [
+        { message: 'Video unavailable', code: VideoUnavailableReasons.Unavailable },
+        { message: 'Private video', code: VideoUnavailableReasons.Private },
+        { message: 'Postprocessing:', code: VideoUnavailableReasons.PostprocessingError },
+        { message: 'The downloaded file is empty', code: VideoUnavailableReasons.EmptyDownload },
+        { message: 'This video is private', code: VideoUnavailableReasons.Private },
+        { message: 'removed by the uploader', code: VideoUnavailableReasons.Private },
+        { message: 'size cap for historical videos', code: VideoUnavailableReasons.Skipped },
       ]
 
       let matchedError = errors.find((e) => errorMsg.includes(e.message))
       if (matchedError) {
-        await this.dynamodbService.videos.updateState(video, 'VideoUnavailable')
-        this.logger.error(`${errorMsg}. Skipping from syncing...`, { videoId: video.id })
+        await this.dynamodbService.videos.updateState(video, `VideoUnavailable::${matchedError.code}`)
+        this.logger.error(`${errorMsg}. Skipping from syncing...`, { videoId: video.id, reason: matchedError.code })
+      }
+
+      // If the error is 403 Forbidden means the IP address was blocked,
+      // restart the proxy server EC2 instance if enabled (i.e setup exists)
+      if (errorMsg.includes('HTTP Error 403: Forbidden') && this.config.chiselProxy?.ec2AutoRotateIp) {
+        await restartEC2Instance(this.logger)
       }
 
       await this.removeVideoFile(video.id)